% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_jmaghirang.R
\name{fars_read}
\alias{fars_read}
\title{Read fars data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{is a string representing the name of the data file}
}
\value{
data frame named data

if filename input is not found, returns error message file does not exist
}
\description{
Read fars data file
}
\examples{
fars_read("accident_2013.csv.bz2")
}
